♻️  Refactor: 포트폴리오 카드 컴포넌트 분리
